refactor: change effect operation arity mismatch to ResolutionError
